+++ changelog.md
@@ -1,3 +1,23 @@
+## Elemental 6.1.2 (2021-03-26)
+
+### Engine
+- Make max element name length 240
+
+## Elemental 6.1.1 (2021-03-22)
+
+### Engine
+- Fixed animations I broke in the last update
+
+## Elemental 6.1.0 (2021-03-22)
+
+### Engine
+- MUCH faster load times with lots of elements
+
+## Elemental 6.0.0 (2021-03-14)
+
+### First Elemental 6!
+- Elemental 6 is now a fork of Elemental 4!
+
 ## Elemental 4.4.1 (2021-02-15)
 
 ### i am stupid.